role controller showing null

+++ src/main/java/com/ms4/client/entity/Role.java
@@ -6,15 +6,16 @@ import javax.persistence.GenerationType;
 import javax.persistence.Id;
 
 import lombok.Data;
+import lombok.Getter;
+import lombok.Setter;
 
 
 @Entity
 @Data
-
 public class Role {
 @GeneratedValue(strategy = GenerationType.IDENTITY)
 @Id
-private Integer id;
+private Long id;
 private String description;
 private String details;
